Move the decoder to a submodule
